+++ pdf22md/Package.swift
@@ -50,4 +50,4 @@ let package = Package(
             ]
         )
     ]
-) 
+)